Move to separate translation units
