+++ dGame/dComponents/ModelComponent.cpp
@@ -7,356 +7,6 @@
 #include "BehaviorStates.h"
 #include "ControlBehaviorMsgs.h"
 
-///////////////// Strip ///////////////////
-
-template<>
-void Strip::HandleMsg(AddStripMessage& msg) {
-	m_Actions = msg.GetActionsToAdd();
-	m_Position = msg.GetPosition();
-};
-
-template<>
-void Strip::HandleMsg(AddActionMessage& msg) {
-	if (msg.GetActionIndex() == -1) return;
-
-	m_Actions.insert(m_Actions.begin() + msg.GetActionIndex(), msg.GetAction());
-};
-
-template<>
-void Strip::HandleMsg(UpdateStripUiMessage& msg) {
-	m_Position = msg.GetPosition();
-};
-
-template<>
-void Strip::HandleMsg(RemoveStripMessage& msg) {
-	m_Actions.clear();
-};
-
-template<>
-void Strip::HandleMsg(RemoveActionsMessage& msg) {
-	if (msg.GetActionIndex() >= m_Actions.size()) return;
-	m_Actions.erase(m_Actions.begin() + msg.GetActionIndex(), m_Actions.end());
-};
-
-template<>
-void Strip::HandleMsg(UpdateActionMessage& msg) {
-	if (msg.GetActionIndex() >= m_Actions.size()) return;
-	m_Actions.at(msg.GetActionIndex()) = msg.GetAction();
-};
-
-template<>
-void Strip::HandleMsg(RearrangeStripMessage& msg) {
-	if (msg.GetDstActionIndex() >= m_Actions.size() || msg.GetSrcActionIndex() >= m_Actions.size() || msg.GetSrcActionIndex() <= msg.GetDstActionIndex()) return;
-	std::rotate(m_Actions.begin() + msg.GetDstActionIndex(), m_Actions.begin() + msg.GetSrcActionIndex(), m_Actions.end());
-};
-
-template<>
-void Strip::HandleMsg(SplitStripMessage& msg) {
-	if (msg.GetTransferredActions().empty() && !m_Actions.empty()) {
-		auto startToMove = m_Actions.begin() + msg.GetSrcActionIndex();
-		msg.SetTransferredActions(startToMove, m_Actions.end());
-		m_Actions.erase(startToMove, m_Actions.end());
-	} else {
-		m_Actions = msg.GetTransferredActions();
-		m_Position = msg.GetPosition();
-	}
-};
-
-template<>
-void Strip::HandleMsg(MergeStripsMessage& msg) {
-	if (msg.GetMigratedActions().empty() && !m_Actions.empty()) {
-		msg.SetMigratedActions(m_Actions.begin(), m_Actions.end());
-		m_Actions.erase(m_Actions.begin(), m_Actions.end());
-	} else {
-		m_Actions.insert(m_Actions.begin() + msg.GetDstActionIndex(), msg.GetMigratedActions().begin(), msg.GetMigratedActions().end());
-	}
-};
-
-template<>
-void Strip::HandleMsg(MigrateActionsMessage& msg) {
-	if (msg.GetMigratedActions().empty() && !m_Actions.empty()) {
-		auto startToMove = m_Actions.begin() + msg.GetSrcActionIndex();
-		msg.SetMigratedActions(startToMove, m_Actions.end());
-		m_Actions.erase(startToMove, m_Actions.end());
-	} else {
-		m_Actions.insert(m_Actions.begin() + msg.GetDstActionIndex(), msg.GetMigratedActions().begin(), msg.GetMigratedActions().end());
-	}
-};
-
-void Strip::SendBehaviorBlocksToClient(AMFArrayValue& args) const {
-	m_Position.SendBehaviorBlocksToClient(args);
-
-	auto* actions = args.InsertArray("actions");
-	for (auto& action : m_Actions) {
-		action.SendBehaviorBlocksToClient(*actions);
-	}
-};
-
-///////////////// State ///////////////////
-
-template<>
-void State::HandleMsg(AddStripMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		m_Strips.resize(msg.GetActionContext().GetStripId() + 1);
-	}
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(AddActionMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		return;
-	}
-
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(UpdateStripUiMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		return;
-	}
-
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(RemoveActionsMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		return;
-	}
-
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(RearrangeStripMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		return;
-	}
-
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(UpdateActionMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		return;
-	}
-
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(RemoveStripMessage& msg) {
-	if (m_Strips.size() <= msg.GetActionContext().GetStripId()) {
-		return;
-	}
-
-	m_Strips.at(msg.GetActionContext().GetStripId()).HandleMsg(msg);
-};
-
-template<>
-void State::HandleMsg(SplitStripMessage& msg) {
-	if (msg.GetTransferredActions().empty()) {
-		if (m_Strips.size() <= msg.GetSourceActionContext().GetStripId()) {
-			return;
-		}
-
-		m_Strips.at(msg.GetSourceActionContext().GetStripId()).HandleMsg(msg);
-	} else {
-		if (m_Strips.size() <= msg.GetDestinationActionContext().GetStripId()) {
-			m_Strips.resize(msg.GetDestinationActionContext().GetStripId() + 1);
-		}
-
-		m_Strips.at(msg.GetDestinationActionContext().GetStripId()).HandleMsg(msg);
-	}
-};
-
-template<>
-void State::HandleMsg(MergeStripsMessage& msg) {
-	if (msg.GetMigratedActions().empty()) {
-		if (m_Strips.size() <= msg.GetSourceActionContext().GetStripId()) {
-			return;
-		}
-
-		m_Strips.at(msg.GetSourceActionContext().GetStripId()).HandleMsg(msg);
-	} else {
-		if (m_Strips.size() <= msg.GetDestinationActionContext().GetStripId()) {
-			m_Strips.resize(msg.GetDestinationActionContext().GetStripId() + 1);
-		}
-
-		m_Strips.at(msg.GetDestinationActionContext().GetStripId()).HandleMsg(msg);
-	}
-};
-
-template<>
-void State::HandleMsg(MigrateActionsMessage& msg) {
-	if (msg.GetMigratedActions().empty()) {
-		if (m_Strips.size() <= msg.GetSourceActionContext().GetStripId()) {
-			return;
-		}
-
-		m_Strips.at(msg.GetSourceActionContext().GetStripId()).HandleMsg(msg);
-	} else {
-		if (m_Strips.size() <= msg.GetDestinationActionContext().GetStripId()) {
-			m_Strips.resize(msg.GetDestinationActionContext().GetStripId() + 1);
-		}
-
-		m_Strips.at(msg.GetDestinationActionContext().GetStripId()).HandleMsg(msg);
-	}
-};
-
-bool State::IsEmpty() const {
-	for (auto& strip : m_Strips) {
-		if (!strip.IsEmpty()) return false;
-	}
-	return true;
-}
-
-void State::SendBehaviorBlocksToClient(AMFArrayValue& args) const {
-	auto* strips = args.InsertArray("strips");
-	for (int32_t stripId = 0; stripId < m_Strips.size(); stripId++) {
-		auto& strip = m_Strips.at(stripId);
-		if (strip.IsEmpty()) continue;
-
-		auto* stripArgs = strips->PushArray();
-		stripArgs->Insert("id", static_cast<double>(stripId));
-
-		strip.SendBehaviorBlocksToClient(*stripArgs);
-	}
-};
-
-///////////////// PropertyBehavior ///////////////////
-
-PropertyBehavior::PropertyBehavior() {
-	m_LastEditedState = BehaviorState::HOME_STATE;
-}
-
-template<>
-void PropertyBehavior::HandleMsg(AddStripMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(AddActionMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(RearrangeStripMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(UpdateActionMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(UpdateStripUiMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(RemoveStripMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(RemoveActionsMessage& msg) {
-	m_States[msg.GetActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(SplitStripMessage& msg) {
-	m_States[msg.GetSourceActionContext().GetStateId()].HandleMsg(msg);
-	m_States[msg.GetDestinationActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetDestinationActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(MigrateActionsMessage& msg) {
-	m_States[msg.GetSourceActionContext().GetStateId()].HandleMsg(msg);
-	m_States[msg.GetDestinationActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetDestinationActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(MergeStripsMessage& msg) {
-	m_States[msg.GetSourceActionContext().GetStateId()].HandleMsg(msg);
-	m_States[msg.GetDestinationActionContext().GetStateId()].HandleMsg(msg);
-	m_LastEditedState = msg.GetDestinationActionContext().GetStateId();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(RenameMessage& msg) {
-	m_Name = msg.GetName();
-};
-
-template<>
-void PropertyBehavior::HandleMsg(AddMessage& msg) {
-	// TODO Parse the corresponding behavior xml file.
-	m_BehaviorId = msg.GetBehaviorId();
-	isLoot = m_BehaviorId != 7965;
-};
-
-void PropertyBehavior::SetBehaviorId(int32_t behaviorId) {
-	m_BehaviorId = behaviorId;
-}
-
-void PropertyBehavior::SendBehaviorListToClient(AMFArrayValue& args) const {
-	args.Insert("id", std::to_string(m_BehaviorId));
-	args.Insert("name", m_Name);
-	args.Insert("isLocked", isLocked);
-	args.Insert("isLoot", isLoot);
-}
-
-void PropertyBehavior::VerifyLastEditedState() {
-	if (!m_States[m_LastEditedState].IsEmpty()) return;
-
-	for (const auto& [stateId, state] : m_States) {
-		if (state.IsEmpty()) continue;
-
-		LOG_DEBUG("Updating last edited state to %i because %i is empty.", stateId, m_LastEditedState);
-		m_LastEditedState = stateId;
-		return;
-	}
-
-	LOG_DEBUG("No states found, sending default state");
-
-	m_LastEditedState = BehaviorState::HOME_STATE;
-}
-
-void PropertyBehavior::SendBehaviorBlocksToClient(AMFArrayValue& args) const {
-	auto* stateArray = args.InsertArray("states");
-
-	auto lastState = BehaviorState::HOME_STATE;
-	for (auto& [stateId, state] : m_States) {
-		if (state.IsEmpty()) continue;
-
-		LOG_DEBUG("Serializing state %i", stateId);
-		auto* stateArgs = stateArray->PushArray();
-		stateArgs->Insert("id", static_cast<double>(stateId));
-		state.SendBehaviorBlocksToClient(*stateArgs);
-	}
-
-	auto* executionState = args.InsertArray("executionState");
-	executionState->Insert("stateID", static_cast<double>(m_LastEditedState));
-	executionState->InsertArray("strips");
-
-	// TODO Serialize the execution state of the behavior
-}
-
-//////////////////////// ModelComponent ////////////////////////
-
 ModelComponent::ModelComponent(Entity* parent) : Component(parent) {
 	m_OriginalPosition = m_Parent->GetDefaultPosition();
 	m_OriginalRotation = m_Parent->GetDefaultRotation();
